feat: enhance profile controller with error handling for get, set, update, and delete operations

+++ server/src/controllers/profileController.js
@@ -3,7 +3,13 @@ import Profile from "../models/Profile.js";
 export const getProfileDetails = async (req, res) => {
   try {
     const profile = await Profile.findOne({ email: req.query.email });
-    res.send(profile);
+    if (profile) {
+      res.send(profile);
+    } else {
+      res.status(404).send({
+        error: "Profile not found!",
+      });
+    }
   } catch (error) {
     res.status(500).send({
       error: "Internal server error!",
@@ -13,12 +19,19 @@ export const getProfileDetails = async (req, res) => {
 
 export const setProfileDetails = async (req, res) => {
   try {
-    const profile = new Profile(req.body);
-    await profile.save();
-    res.send({
-      message: "Profile created successfully!",
-      data: profile,
-    });
+    const profile = await Profile.findOne({ email: req.query.email });
+    if (!profile) {
+      const profile = new Profile(req.body);
+      await profile.save();
+      res.send({
+        message: "Profile created successfully!",
+        data: profile,
+      });
+    } else {
+      res.status(403).send({
+        error: "Profile already exists!",
+      });
+    }
   } catch (error) {
     res.status(500).send({
       error: "Internal server error!",
@@ -29,12 +42,18 @@ export const setProfileDetails = async (req, res) => {
 export const updateProfileDetails = async (req, res) => {
   try {
     const profile = await Profile.findOne({ email: req.body.email });
-    await profile.updateOne(req.body);
-    profile.save();
-    res.send({
-      message: "Profile updated successfully!",
-      data: profile,
-    });
+    if (profile) {
+      await profile.updateOne(req.body);
+      await profile.save();
+      res.send({
+        message: "Profile updated successfully!",
+        data: profile,
+      });
+    } else {
+      res.status(404).send({
+        error: "Profile not found!",
+      });
+    }
   } catch (error) {
     res.status(500).send({
       error: "Internal server error!",
@@ -45,14 +64,21 @@ export const updateProfileDetails = async (req, res) => {
 export const deleteProfileDetails = async (req, res) => {
   try {
     const profile = await Profile.findOne({ email: req.query.email });
-    await profile.deleteOne(profile)
-    res.send({
-      message: "Profile deleted successfully!",
-      data: profile,
-    });
+    if(profile) {
+      await profile.deleteOne(profile);
+      res.send({
+        message: "Profile deleted successfully!",
+        data: profile,
+      });
+    } else {
+      res.status(404).send({
+        error: "Profile not found!",
+      });
+    }
+
   } catch (error) {
     res.status(500).send({
       error: "Internal server error!",
     });
   }
-}
+};
